add message success and fix an error in the expenseResolver (=! null to == null
